requirements.txt
astroid==2.2.5
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.5.0
Flask-Migrate==2.7.0
python-jose-cryptodome==1.3.2
Flask-Cors==3.0.8
